requirements.txt
google-generativeai>=0.5.0
SpeechRecognition>=3.10.0
gTTS>=2.5.0
pygame>=2.5.0
streamlit>=1.35.0
pyaudio>=0.2.14